review: keep one reference splitter for both engines
The containerd engine dropped the tag it never used and the process engine kept it; one shape in engine/types serves both, with the test.
