+++ engine/containerd/utils.go
@@ -1,8 +1,6 @@
 package containerd
 
 import (
-	"strings"
-
 	"github.com/distribution/reference"
 )
 
@@ -12,13 +10,4 @@ func normalizeRef(ref string) string {
 		return ref
 	}
 	return named.String()
-}
-
-// imageName drops an image reference's tag, ignoring a registry port.
-func imageName(ref string) string {
-	colon := strings.LastIndex(ref, ":")
-	if colon < 0 || colon < strings.LastIndex(ref, "/") {
-		return ref
-	}
-	return ref[:colon]
 }
